+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Add `:classpath` to `serverInfo` command for downstream usages.
   - Fix issues in `extract-function` that arise when operating over `.cljc` files.
   - Add setting `:copy-kondo-configs?` to allow disable auto clj-kondo config copy, enabled by default. #694
+  - Lint unused public keywords, like unused re-frame events/subs. #691
 
 - Editor
   - Improve 'create function' refactor code action handling multiple cases. #682